refactor: simplify fits validation
